// src/Interfaces/Workflows/ICallToAction.ts
 export interface ICallToAction {
    title: string | undefined;
    styles: string | undefined;
    contextType: string | undefined;
    contextIdVariable: string | undefined;
    command: string | undefined;
    commandConfig: string | undefined;
    icon: string | undefined;
    css: string | undefined;

}